[FIXED] - some issues

+++ OpenGLSetup/source/graphics/Camera.cpp
@@ -1,5 +1,4 @@
 #include "Pch.h"
-#include "GLFW/glfw3.h"
 #include "graphics/Camera.h"
 #include <gtc/matrix_transform.hpp>
 
@@ -11,19 +10,6 @@ glm::mat4 Camera::GetProjectionMatrix(float screenWidth, float inverseScreenHeig
 {
     return glm::perspective(glm::radians(m_fov), screenWidth * inverseScreenHeight, 0.1f, 100.0f);
 }
-
-//void Camera::ProcessKeyboardInput(GLFWwindow*& window, const float deltaTime)
-//{
-//    float velocity = m_movementSpeed * deltaTime;
-//    if (glfwGetKey(window, GLFW_KEY_W) == GLFW_PRESS)
-//        m_position += m_front * velocity;
-//    if (glfwGetKey(window, GLFW_KEY_S) == GLFW_PRESS)
-//        m_position -= m_front * velocity;
-//    if (glfwGetKey(window, GLFW_KEY_A) == GLFW_PRESS)
-//        m_position -= m_right * velocity;
-//    if (glfwGetKey(window, GLFW_KEY_D) == GLFW_PRESS)
-//        m_position += m_right * velocity;
-//}
 
 void Camera::ProcessMouseInput(float xOffset, float yOffset, const bool constrainPitch)
 {
@@ -53,9 +39,9 @@ Camera::~Camera()
 void Camera::UpdateVectors()
 {
     glm::vec3 front{};
-    front.x = cosf(glm::radians(m_yaw)) * cos(glm::radians(m_pitch));
+    front.x = cosf(glm::radians(m_yaw)) * cosf(glm::radians(m_pitch));
     front.y = sinf(glm::radians(m_pitch));
-    front.z = sinf(glm::radians(m_yaw)) * cos(glm::radians(m_pitch));
+    front.z = sinf(glm::radians(m_yaw)) * cosf(glm::radians(m_pitch));
     m_front = glm::normalize(front);
     m_right = glm::normalize(glm::cross(m_front, m_worldUp));
     m_up = glm::normalize(glm::cross(m_right, m_front));
